Updated ctrlp.vim
There has been a problem where the first time I checked out the
repository for ctrlp.vim I got the wrong repository, then removed it
incorrectly and cloned on top of that. In some way the submodule is
corrupted and it has the commit from the other repository and then it
tries to fetch the commit from the real repository and it cannot find a
link between both commits because they belong to different repositories.
This should be fixed by fetching from the correct repository and
committing, if not possible, then removing the submodule, committing and
then re-adding the submodule should work.