Added a text template and fixed settings
Added a folder with text templates that can be inserted into the code
and a few commands to insert different text snippets. Added shellslash
in order to be able to use more or less general commands for windows and
linux when setting up backups, swaps and undo. I can probably still
shorten the configuration file by merging the setting of the directories
since the creation is already merged. Added a new mapping to get me the
full name of the file I'm currently working on with F4.